Implemented user permissions + fixed bugs + added routes for CRUD

+++ back-end/src/post/api/views.py
@@ -6,13 +6,16 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 
 from account.models import Account
+from post.models import Post
 # Import our Registration Serializer
 from account.api.serializers import RegistrationSerializer
+from post.api.serializers import PostSerializer
 
 from rest_framework.authtoken.models import Token
 
 # This method is called when a user uses the sign up form
 @api_view(['GET'])
+@permission_classes((IsAuthenticated,))
 def get_post_view(request, id):
     try:
         post = Post.objects.get(id=id)
@@ -25,11 +28,16 @@ def get_post_view(request, id):
 
 
 @api_view(['PUT'])
+@permission_classes((IsAuthenticated,))
 def update_post_view(request, id):
     try:
         post = Post.objects.get(id=id)
     except Post.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    account = request.user
+    if post.account != account:
+        return Response({'response': "You don't have permission to edit"})
 
     if request.method == "PUT":
         serializer = PostSerializer(post, data=request.data)
@@ -43,11 +51,16 @@ def update_post_view(request, id):
 
 
 @api_view(['DELETE'])
+@permission_classes((IsAuthenticated,))
 def delete_post_view(request, id):
     try:
         post = Post.objects.get(id=id)
     except Post.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    account = request.user
+    if post.account != account:
+        return Response({'response': "You don't have permission to delete"})
 
     if request.method == "DELETE":
         operation = post.delete()
@@ -60,8 +73,9 @@ def delete_post_view(request, id):
 
 
 @api_view(['POST'])
+@permission_classes((IsAuthenticated,))
 def create_post_view(request):
-    account = Account.objects.get(pk=1)
+    account = request.user
 
     post = Post(account=account)
 
@@ -70,5 +84,5 @@ def create_post_view(request):
         data = {}
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATE)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
